Feat : post-api 완성

+++ src/pages/post/Post.js
@@ -14,12 +14,13 @@ const Div = styled.div`
   position: relative;
   width: 100vw;
 `
-const userData = JSON.parse(localStorage.getItem('user'))
-const id = 225
+
 const LIMIT = 4
 
 const Post = () => {
-  const [userId] = useState(id)
+  const userData = JSON.parse(localStorage.getItem('user'))
+  const id = userData.id
+
   const [isOpened, setIsOpened] = useState(false)
   const [cnt, setCnt] = useState(0)
   const [items, setItems] = useState([])
@@ -28,15 +29,14 @@ const Post = () => {
   const [isDeleteLoading, isDeleteError, postMainDeleteAsync] =
     useAsync(postMainDelete)
   const count = items.length
+
   const state =
     useLocation().pathname.split('/').length === 4 ? 'answer' : 'default'
   const navigate = useNavigate()
-  const location = useLocation()
 
-  const handleLoad = async (options) => {
+  const handleLoad = async (options = { id: id, offset: 0, limit: LIMIT }) => {
     const result = await postMainDataAsync(options)
     if (!result) return
-    console.log(result)
     const { results } = result
     setCnt(result.count)
     if (options.offset === 0) {
@@ -54,12 +54,16 @@ const Post = () => {
   }
 
   const handleDeleteButton = async (id) => {
-    const result = await postMainDeleteAsync(id)
-    if (!result) return
-    window.localStorage.clear()
-    if (isDeleteError) return <div>에러!</div>
-    if (isDeleteLoading) return <div>로딩중!</div>
-    navigate('/')
+    if (window.confirm('삭제 하시겠습니까?')) {
+      const result = await postMainDeleteAsync(id)
+      if (!result) return
+      window.localStorage.clear()
+      if (isDeleteError) return <div>에러!</div>
+      if (isDeleteLoading) return <div>로딩중!</div>
+      navigate('/')
+    } else {
+      return
+    }
   }
 
   useEffect(() => {
@@ -71,16 +75,14 @@ const Post = () => {
   useEffect(() => {
     handleLoad({ id, offset: 0, limit: LIMIT })
   }, [])
-  // console.log(items)
+
   return (
     <>
       <Div>
-        <Nav userData={location.state || userData} />
-        <S.Div className="Div">
+        <Nav userData={userData} />
+        <S.Div>
           {state === 'answer' && (
-            <S.DeleteButton
-              onClick={() => handleDeleteButton(location.state?.id || userId)}
-            >
+            <S.DeleteButton onClick={() => handleDeleteButton(id)}>
               <PostDeleteButton />
             </S.DeleteButton>
           )}
@@ -88,11 +90,11 @@ const Post = () => {
           {count !== 0 ? (
             <PostContent
               setIsOpened={setIsOpened}
+              isOpened={isOpened}
               state={state}
               items={items}
-              isOpened={isOpened}
-              handleLoadMore={handelLoadMore}
               cnt={cnt}
+              handleLoadMore={handelLoadMore}
             />
           ) : (
             <PostNoContent
